+++ internal/plugins/jest/all.go
@@ -2,7 +2,7 @@ package jest
 
 import (
 	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/expect_expect"
-	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/max_expect"
+	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/max_expects"
 	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/max_nested_describe"
 	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/no_alias_methods"
 	"github.com/web-infra-dev/rslint/internal/plugins/jest/rules/no_commented_out_tests"
@@ -36,7 +36,7 @@ import (
 func GetAllRules() []rule.Rule {
 	return []rule.Rule{
 		expect_expect.ExpectExpectRule,
-		max_expect.MaxExpectRule,
+		max_expects.MaxExpectsRule,
 		max_nested_describe.MaxNestedDescribeRule,
 		no_alias_methods.NoAliasMethodsRule,
 		no_commented_out_tests.NoCommentedOutTestsRule,
