+++ internal/plugins/jest/rules/max_expects/max_expects.md
